+++ src/main/java/io/tickerstorm/data/YahooRealtimeQuoteQuery.java
@@ -46,7 +46,7 @@ public class YahooRealtimeQuoteQuery implements QueryBuilder, DataConverter {
     else
       return null;
 
-    c.askSize = new BigInteger(data[2]);
+    c.askSize = new BigDecimal(data[2]);
 
     if (!data[3].equalsIgnoreCase("N/A"))
       c.bid = new BigDecimal(data[3]);
@@ -55,7 +55,7 @@ public class YahooRealtimeQuoteQuery implements QueryBuilder, DataConverter {
     else
       return null;
 
-    c.bidSize = new BigInteger(data[5]);
+    c.bidSize = new BigDecimal(data[5]);
     c.exchange = data[6];
 
     return new Quote[] { c };
